+++ server/models/ordersTable.js
@@ -5,9 +5,9 @@ export async function createOrdersTable() {
             CREATE TABLE IF NOT EXISTS orders ( 
                 id UUID DEFAULT gen_random_uuid() PRIMARY KEY,
                 buyer_id UUID NOT NULL, 
-                total_price INT NOT NULL CHECK (total_price > 0), 
-                tax_price INT NOT NULL CHECK (tax_price >= 0), 
-                shipping_price INT NOT NULL CHECK (shipping_price >= 0), 
+                total_price NUMERIC(10,2) NOT NULL CHECK (total_price > 0), 
+                tax_price NUMERIC(5,2) NOT NULL CHECK (tax_price >= 0), 
+                shipping_price NUMERIC(5,2) NOT NULL CHECK (shipping_price >= 0), 
                 order_status VARCHAR(50) DEFAULT 'Processing' CHECK (order_status IN ('Processing', 'Shipped', 'Delivered', 'Cancelled')), 
                 paid_at TIMESTAMP CHECK (paid_at IS NULL OR paid_at <= CURRENT_TIMESTAMP), 
                 created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
